+++ src/main/java/icu/samnyan/aqua/sega/diva/model/request/card/ChangePasswdRequest.java
@@ -15,9 +15,9 @@ import lombok.Setter;
 @NoArgsConstructor
 public class ChangePasswdRequest extends BaseRequest {
     private String a_code;
-    private int aime_id;
+    private long aime_id;
     private String aime_a_code;
-    private int pd_id;
+    private long pd_id;
     private int accept_idx;
     private String new_passwd;
 }
