+++ src/main/java/icu/samnyan/aqua/sega/diva/model/request/card/RegistrationRequest.java
@@ -19,7 +19,7 @@ public class RegistrationRequest extends BaseRequest {
     private String mmgameid;
     private String mmuid;
     private String a_code;
-    private int aime_id;
+    private long aime_id;
     private String aime_a_code;
     private String key_obj_type;
     private String player_name;
