+++ src/main/java/icu/samnyan/aqua/sega/diva/model/request/user/PreStartRequest.java
@@ -19,7 +19,7 @@ public class PreStartRequest extends BaseRequest {
     private String mmgameid;
     private String mmuid;
     private String a_code;
-    private int aime_id;
+    private long aime_id;
     private String aime_a_code;
     private String key_obj_type;
     private boolean exec_vu;
